Refactoring of MovieListInteractor, creation of an data manager that is responsible for getting movie list and genre list and combining it.
